requirements.txt
pyyaml
numpy
opencv-python
pandas
matplotlib
seaborn